+++ core/src/main/scala/com/krux/hyperion/HyperionCli.scala
@@ -6,7 +6,7 @@ import com.krux.hyperion.cli.EntryPoint
   * HyperionCli is a base trait that brings in CLI functionality to
   * DataPipelineDef's.
   */
-trait HyperionCli { this: DataPipelineDef =>
+trait HyperionCli { this: DataPipelineDefGroup =>
 
   def main(args: Array[String]): Unit = System.exit(EntryPoint(this).run(args))