+++ core/src/main/scala/com/krux/hyperion/activity/SparkStep.scala
@@ -1,7 +1,7 @@
 package com.krux.hyperion.activity
 
 import com.krux.hyperion.adt.{ HS3Uri, HString, HInt }
-import com.krux.hyperion.common.Memory
+import com.krux.hyperion.common.{ Memory, Escapable }
 import com.krux.hyperion.HyperionContext
 
 /**
@@ -41,7 +41,7 @@ case class SparkStep private (
     Seq(jarUri.serialize),
     mainClass.map(_.toString).toSeq,
     args
-  ).flatten.mkString(",")
+  ).flatten.map(x => Escapable.escape(x.toString, ',')).mkString(",")
 
   override def toString: String = serialize
 }