+++ core/src/main/scala/com/krux/hyperion/cli/Action.scala
@@ -1,7 +1,7 @@
 package com.krux.hyperion.cli
 
-import com.krux.hyperion.DataPipelineDef
+import com.krux.hyperion.DataPipelineDefGroup
 
 private[hyperion] trait Action {
-  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean
+  def apply(options: Options, pipelineDef: DataPipelineDefGroup): Boolean
 }