+++ core/src/main/scala/com/krux/hyperion/cli/ActivateAction.scala
@@ -1,7 +1,8 @@
 package com.krux.hyperion.cli
 
-import com.krux.hyperion.HyperionAwsClient
+import com.krux.hyperion.client.AwsClientForDef
 
 private[hyperion] case object ActivateAction extends AwsAction {
-  def apply(options: Options, client: HyperionAwsClient): Boolean = client.activatePipeline()
+  def apply(options: Options, client: AwsClientForDef): Boolean =
+    client.forName().flatMap(_.forId()).flatMap(_.activatePipelines()).isDefined
 }