+++ core/src/main/scala/com/krux/hyperion/cli/DeleteAction.scala
@@ -1,7 +1,10 @@
 package com.krux.hyperion.cli
 
-import com.krux.hyperion.HyperionAwsClient
+import com.krux.hyperion.client.AwsClientForDef
 
 private[hyperion] case object DeleteAction extends AwsAction {
-  def apply(options: Options, client: HyperionAwsClient): Boolean = client.deletePipeline()
+
+  def apply(options: Options, client: AwsClientForDef): Boolean =
+    client.forName().flatMap(_.forId()).flatMap(_.deletePipelines()).isDefined
+
 }