+++ core/src/main/scala/com/krux/hyperion/cli/GenerateAction.scala
@@ -2,12 +2,23 @@ package com.krux.hyperion.cli
 
 import java.io.PrintStream
 
-import com.krux.hyperion.DataPipelineDef
 import org.json4s.jackson.JsonMethods._
 
+import com.krux.hyperion.DataPipelineDefGroup
+
+
 private[hyperion] case object GenerateAction extends Action {
-  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean = {
-    options.output.map(f => new PrintStream(f)).getOrElse(System.out).println(pretty(render(pipelineDef)))
+
+  def apply(options: Options, defGroup: DataPipelineDefGroup): Boolean = {
+
+    defGroup.ungroup().foreach { case (key, pipelineDef) =>
+      val outputStream = options.output
+        .map(o => new PrintStream(o + key.map(pipelineDef.nameKeySeparator + _).getOrElse("")))
+        .getOrElse(System.out)
+      outputStream.println(pretty(render(pipelineDef)))
+    }
+
     true
   }
+
 }