+++ core/src/main/scala/com/krux/hyperion/cli/GraphAction.scala
@@ -2,14 +2,23 @@ package com.krux.hyperion.cli
 
 import java.io.PrintStream
 
-import com.krux.hyperion.DataPipelineDef
+import com.krux.hyperion.DataPipelineDefGroup
 import com.krux.hyperion.workflow.WorkflowGraphRenderer
 
 private[hyperion] case object GraphAction extends Action {
-  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean = {
-    val renderer = WorkflowGraphRenderer(pipelineDef, options.removeLastNameSegment,
-      options.label, options.includeResources, options.includeDataNodes, options.includeDatabases)
-    options.output.map(f => new PrintStream(f)).getOrElse(System.out).println(renderer.render())
+
+  def apply(options: Options, defGroup: DataPipelineDefGroup): Boolean = {
+
+    defGroup.ungroup().foreach { case (key, pipelineDef) =>
+      val renderer = WorkflowGraphRenderer(pipelineDef, options.removeLastNameSegment,
+        options.label, options.includeResources, options.includeDataNodes, options.includeDatabases)
+      options.output
+        .map(f => new PrintStream(f + key.map(pipelineDef.nameKeySeparator + _).getOrElse("") + ".dot"))
+        .getOrElse(System.out)
+        .println(renderer.render())
+    }
     true
+
   }
+
 }