+++ core/src/main/scala/com/krux/hyperion/package.scala
@@ -1,0 +1,9 @@
+package com.krux
+
+package object hyperion {
+
+  type WorkflowKey = Option[String]
+
+  final val EmptyKey = None
+
+}