+++ examples/src/test/scala/com/krux/hyperion/examples/ExampleRedshiftLoadSpec.scala
@@ -4,6 +4,7 @@ import org.scalatest.WordSpec
 import org.json4s.jackson.JsonMethods._
 import org.json4s.JsonDSL._
 import org.json4s._
+
 import com.krux.hyperion.DataPipelineDef._
 
 
@@ -38,7 +39,7 @@ class ExampleRedshiftLoadSpec extends WordSpec {
         ("id" -> ec2Id) ~
         ("name" -> ec2Id) ~
         ("terminateAfter" -> "8 hours") ~
-        ("imageId" -> "ami-b0682cd8") ~
+        ("imageId" -> "ami-0188776c") ~
         ("instanceType" -> "m1.small") ~
         ("region" -> "us-east-1") ~
         ("securityGroups" -> Seq("your-security-group")) ~